[refactor] QuizServiceImpl.java 삭제
